Extract TOML parsing as a feature
