+++ .gitignore
@@ -2,4 +2,3 @@ target/
 **/*.rs.bk
 Cargo.lock
 *.swp
-config.rs
