+++ config_struct/src/toml_parsing.rs
@@ -1,0 +1,71 @@
+use std::path::Path;
+
+use toml::{ self, Value };
+
+use { RawValue, RawStructValue, create_config_module };
+
+
+pub fn create_module_from_config<S, D>(config_path: S, module_path: D)
+where
+    S: AsRef<Path>,
+    D: AsRef<Path>
+{
+    use std::collections::BTreeMap;
+    use std::fs::File;
+    use std::io::{ Read, Write };
+
+    let config_source = {
+        let mut buffer = String::new();
+        let file = &mut File::open(&config_path).unwrap();
+        file.read_to_string(&mut buffer).unwrap();
+        buffer
+    };
+
+    let toml_object: BTreeMap<String, Value> = toml::from_str(&config_source).unwrap();
+
+    let raw_config: RawStructValue = {
+        let struct_name = "Config".to_owned();
+        let fields = toml_object.into_iter().map(|(key, value)|
+        {
+            let value = toml_to_raw_value("_Config", &key, value);
+            (key, value)
+        }).collect();
+        RawStructValue { struct_name, fields }
+    };
+
+    let config_module_source = create_config_module(&raw_config);
+
+    let file = &mut File::create(&module_path).unwrap();
+    file.write_all(config_module_source.as_bytes()).unwrap();
+}
+
+
+fn toml_to_raw_value(super_struct: &str, super_key: &str, value: Value) -> RawValue
+{
+    match value
+    {
+        Value::Boolean(value) => RawValue::Bool(value),
+        Value::Integer(value) => RawValue::I64(value),
+        Value::Float(value) => RawValue::F64(value),
+        Value::String(value) => RawValue::String(value),
+        Value::Datetime(value) => RawValue::String(value.to_string()),
+        Value::Array(values) => {
+            RawValue::Array(values.into_iter()
+                .map(|value| toml_to_raw_value(super_struct, super_key, value))
+                .collect())
+        },
+        Value::Table(values) => {
+            let sub_struct_name = format!("{}__{}", super_struct, super_key);
+            let values = values.into_iter()
+                .map(
+                    |(key, value)|
+                    {
+                        let value = toml_to_raw_value(&sub_struct_name, &key, value);
+                        (key, value)
+                    })
+                .collect();
+            RawValue::Struct(RawStructValue { struct_name: sub_struct_name, fields: values })
+        }
+    }
+}
+
