+++ config_struct_test/build.rs
@@ -3,5 +3,7 @@ extern crate config_struct;
 
 fn main()
 {
-    config_struct::construct_config("config.toml", "src/config.rs");
+    use config_struct::toml_parsing;
+
+    toml_parsing::create_module_from_config("config.toml", "src/config/toml.rs");
 }
