+++ config_struct_test/src/config/.gitignore
@@ -1,0 +1,2 @@
+*.rs
+!mod.rs
