+++ config_struct_test/src/config/mod.rs
@@ -1,0 +1,2 @@
+pub mod toml;
+
